PoC segmentation GeoJSON
